+++ bin/index.js
@@ -4,6 +4,9 @@ const path = require('path')
 const directoryCommand = require('directory-command')
 
 const directory = path.join(__dirname, '..', 'commands')
-const options = { context: {} }
+const options = {
+  commandName: 'daylog',
+  context: {}
+}
 
 directoryCommand(directory, process.argv.slice(2), options)