+++ commands/index.js
@@ -1,1 +1,7 @@
-module.exports = require('./help')
+function command (args, flags, context) {
+  context.help()
+}
+
+module.exports = {
+  command
+}